% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandbox_script.R
\name{count_authors}
\alias{count_authors}
\title{count number of authors from search}
\usage{
count_authors(df)
}
\arguments{
\item{df}{A dataframe derived from a search. Must have package_data as one
of the columns}
}
\description{
count number of authors from search
}
\author{
Kathryn Doering
}
